refactor: extract nav locations
