+++ src/pages/topics/index.tsx
@@ -6,15 +6,13 @@ import { TopicService } from 'src/domain/topic/service/topicService';
 import { TopicEntity, TopicEntityFactory } from 'src/view/types/topic';
 import { useDispatch } from 'react-redux';
 import { setTopics } from 'src/data/redux/topic/slice';
-import {
-  BottomNavigation,
-  NavLocations,
-} from 'src/components/navbar/bottomNavigation/BottomNavigation';
+import { BottomNavigation } from 'src/components/navbar/bottomNavigation/BottomNavigation';
 import { TopicContainer } from 'src/components/topic/TopicContainer';
 import { pageTitle, rootPath } from 'src/view/route/pagePath';
 import Head from 'next/head';
 import { InvitationCodeForm } from 'src/components/topic/invitation/InvitationCodeForm';
 import { useRouter } from 'next/router';
+import { NavLocations } from 'src/view/types/navigation';
 
 type Props = {
   pickup: TopicEntity | null,
